+++ R/download.R
@@ -63,7 +63,7 @@ download <- function(url, destfile, type = NULL, quiet = FALSE, headers = NULL) 
   on.exit(callback(), add = TRUE)
 
   # form path to temporary file
-  tempfile <- renv_tempfile_create(tmpdir = dirname(destfile))
+  tempfile <- renv_tempfile_path(tmpdir = dirname(destfile))
 
   # request the download
   before <- Sys.time()
@@ -216,7 +216,7 @@ renv_download_default_agent_scope_impl <- function(headers, envir = NULL) {
 
 renv_download_curl <- function(url, destfile, type, request, headers) {
 
-  file <- renv_tempfile_create("renv-download-config-")
+  file <- renv_tempfile_path("renv-download-config-")
 
   fields <- c(
     "user-agent" = renv_http_useragent(),
@@ -330,7 +330,7 @@ renv_download_curl_config <- function() {
 
 renv_download_wget <- function(url, destfile, type, request, headers) {
 
-  config <- renv_tempfile_create("renv-download-config-")
+  config <- renv_tempfile_path("renv-download-config-")
 
   fields <- c(
     "user-agent" = renv_http_useragent(),
@@ -481,7 +481,7 @@ renv_download_headers <- function(url, type, headers) {
     return(list())
 
   # perform the download
-  file <- renv_tempfile_create("renv-headers-")
+  file <- renv_tempfile_path("renv-headers-")
 
   status <- renv_download_impl(
     url      = url,
@@ -629,7 +629,7 @@ renv_download_local <- function(url, destfile, headers) {
     # perform the copy
     before <- Sys.time()
     status <- catch(method(url, destfile, headers))
-    after <- Sys.time()
+    after  <- Sys.time()
 
     # check for success
     if (!identical(status, TRUE))
@@ -664,6 +664,7 @@ renv_download_local_copy <- function(url, destfile, headers) {
   }
 
   # attempt to copy
+  ensure_parent_directory(destfile)
   status <- catchall(renv_file_copy(url, destfile, overwrite = TRUE))
   if (!identical(status, TRUE))
     return(FALSE)