+++ tests/testthat/helper-aaa.R
@@ -13,8 +13,6 @@ test_that <- function(desc, code) {
 
   testthat::skip_if(iscran && renv_platform_macos())
 
-  options(repos = getOption("renv.tests.repos"))
-
   oldlibpaths <- .libPaths()
   oldrepos <- getOption("repos")