+++ tests/testthat/test-hydrate.R
@@ -5,7 +5,7 @@ test_that("hydrate does not change library paths", {
 
   renv_tests_scope()
 
-  lib <- renv_tempfile_create()
+  lib <- renv_tempfile_path()
   ensure_directory(lib)
   .libPaths(lib)