+++ tests/testthat/test-restore.R
@@ -3,7 +3,7 @@ context("Restore")
 
 test_that("library permissions are validated before restore", {
   skip_on_os("windows")
-  inaccessible <- renv_tempfile_create()
+  inaccessible <- renv_tempfile_path()
   dir.create(inaccessible, mode = "0100")
   renv_scope_options(renv.verbose = FALSE)
   expect_false(renv_install_preflight_permissions(inaccessible))
@@ -148,6 +148,8 @@ test_that("renv::restore(packages = <...>) works", {
 })
 
 test_that("restore ignores packages of incompatible architecture", {
+
+  renv_scope_options(renv.tests.verbose = FALSE)
 
   renv_tests_scope(c("unixonly", "windowsonly"))
   renv::init()